[ESIMD] Add more tests for new xmx::dpas() (intel/llvm-test-suite#1291)
* Add the test for bfloat16;
* Add the tests for dg2;
* Run DG2 and PVC tests on esimd_emulator;

+++ SYCL/ESIMD/dpas/dpas_bf16.cpp
@@ -1,4 +1,4 @@
-//==---------------- dpas_fp16.cpp  - DPC++ ESIMD on-device test ----------==//
+//==---------------- dpas_bf16.cpp  - DPC++ ESIMD on-device test ----------==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -10,7 +10,7 @@
 // RUN: %clangxx -fsycl -fsycl-device-code-split=per_kernel %s -o %t.out
 // RUN: %GPU_RUN_PLACEHOLDER %t.out
 
-// This test verifies DPAS support for float16.
+// This test verifies DPAS support for bfloat16.
 
 #include "dpas_common.hpp"
 
@@ -21,13 +21,13 @@ int main(int argc, const char *argv[]) {
   bool Passed = true;
 
   constexpr bool LetDeduceArgs = true;
-  Passed &= tests<8, 8, fp16, fp16, LetDeduceArgs>(Q, Print);
-  Passed &= tests<8, 4, fp16, fp16, LetDeduceArgs>(Q, Print);
-  Passed &= tests<8, 1, fp16, fp16, LetDeduceArgs>(Q, Print);
+  Passed &= tests<8, 8, bf16, bf16, LetDeduceArgs>(Q, Print);
+  Passed &= tests<8, 4, bf16, bf16, LetDeduceArgs>(Q, Print);
+  Passed &= tests<8, 1, bf16, bf16, LetDeduceArgs>(Q, Print);
 
   // TODO: Enable these cases when esimd::simd(ptr) constructor is fixed.
-  // Passed &= tests<8, 5, fp16, fp16, LetDeduceArgs>(Q, Print);
-  // Passed &= tests<8, 3, fp16, fp16, LetDeduceArgs>(Q, Print);
+  // Passed &= tests<8, 5, bf16, bf16, LetDeduceArgs>(Q, Print);
+  // Passed &= tests<8, 3, bf16, bf16, LetDeduceArgs>(Q, Print);
 
   std::cout << (Passed ? "Test Passed\n" : "Test FAILED\n");
   return Passed ? 0 : 1;
